Fix for GPXupload features validation using DragNDrop.
I also moved validateFeatures to the app.utils because it is now used on two places...
